Move mock providers to mocks
