+++ packages/ui/test/hooks/useMyMemberships.test.tsx
@@ -2,7 +2,7 @@ import { expect } from '@jest/globals'
 import { act, renderHook } from '@testing-library/react-hooks'
 import React from 'react'
 import { useMyMemberships } from '../../src/hooks/useMyMemberships'
-import { MockQueryNodeProviders } from '../helpers/providers'
+import { MockQueryNodeProviders } from '../mocks/providers'
 import { alice, bobStash } from '../mocks/keyring'
 import { getMember } from '../mocks/members'
 import { setupMockServer } from '../mocks/server'
