+++ packages/ui/test/modals/InviteMemberModal.test.tsx
@@ -10,7 +10,7 @@ import { Account } from '../../src/common/types'
 import { InviteMemberModal } from '../../src/modals/InviteMemberModal'
 import { ApiContext } from '../../src/providers/api/context'
 import { UseApi } from '../../src/providers/api/provider'
-import { MockKeyringProvider, MockQueryNodeProviders } from '../helpers/providers'
+import { MockKeyringProvider, MockQueryNodeProviders } from '../mocks/providers'
 import { selectMember } from '../helpers/selectMember'
 import { alice, aliceStash } from '../mocks/keyring'
 import { getMember } from '../mocks/members'
