+++ packages/ui/test/modals/TransferInviteModal.test.tsx
@@ -8,7 +8,7 @@ import { TransferIcon } from '../../src/components/icons'
 import { TransferInviteModal } from '../../src/modals/TransferInviteModal'
 import { ApiContext } from '../../src/providers/api/context'
 import { UseApi } from '../../src/providers/api/provider'
-import { MockKeyringProvider, MockQueryNodeProviders } from '../helpers/providers'
+import { MockKeyringProvider, MockQueryNodeProviders } from '../mocks/providers'
 import { selectMember } from '../helpers/selectMember'
 import { getMember } from '../mocks/members'
 import { setupMockServer } from '../mocks/server'
