+++ packages/ui/test/pages/Accounts.test.tsx
@@ -11,7 +11,7 @@ import { Accounts } from '../../src/pages/Profile/MyAccounts/Accounts'
 import { KeyringContext } from '../../src/providers/keyring/context'
 import { MembershipContext } from '../../src/providers/membership/context'
 import { shortenAddress } from '../../src/utils/formatters'
-import { MockApolloProvider } from '../helpers/providers'
+import { MockApolloProvider } from '../mocks/providers'
 import { alice, aliceStash, bob, bobStash } from '../mocks/keyring'
 import { getMember } from '../mocks/members'
 import { setupMockServer } from '../mocks/server'
